requirements.txt
pbr>=1.6
jsonschema>=2.0.0,<3.0.0
alembic>=0.4.1
Babel>=1.3
iso8601>=0.1.9
oauthlib>=0.6
oslo.config>=1.11.0
oslo.context>=0.2.0
oslo.utils>=1.4.0
pecan>=0.4.5
oslo.db>=1.8.0,<4.8.0
oslo.log>=1.0.0
pika>=0.9.14
python-openid
PyYAML>=3.1.0
requests>=1.1
six>=1.7.0
SQLAlchemy>=0.9.7,<=0.9.99
WSME>=0.6,<0.8
sqlalchemy-migrate>=0.9.1,!=0.9.2
SQLAlchemy-FullText-Search>=0.2.3
eventlet>=0.13.0
stevedore>=1.3.0
tzlocal>=1.1.2
email>=4.0.2;python_version<'3.0'
Jinja2>=2.7.3
PyMySQL>=0.6.2,!=0.6.4,<0.7.7
apscheduler>=3.0.1,<3.1.0
python_dateutil>=2.4.0
oslo.concurrency>=3.8.0         # Apache-2.0
oslo.i18n>=2.1.0  # Apache-2.0
